Refactor: Fix compilation errors, fix missing tags in payload, submit state not resetting
