+++ x-pack/platform/plugins/shared/onechat/public/application/components/tools/esql/form/validation/esql_tool_form_validation.ts
@@ -147,6 +147,7 @@ export const esqlFormValidationSchema = z
           }
         });
       }),
+    tags: z.array(z.string()),
   })
   .superRefine(({ esql, params }, ctx) => {
     const inferredParams = extractEsqlParams(esql);
